+++ HardShipAPI/Models/Debt.cs
@@ -1,6 +1,6 @@
 ﻿public class Debt
 {
-    public long DebtID { get; set; }
+    public int DebtID { get; set; }
     public string Name { get; set; }
     public string DOB { get; set; }
     public decimal? Income { get; set; }
