+++ HardShipAPI/Models/Hardship.cs
@@ -14,9 +14,9 @@
     }
     public class Hardship
     {
-        public long HardshipID { get; set; }
+        public int HardshipID { get; set; }
         public short HardshipTypeID { get; set; }
-        public long DebtID { get; set; }
+        public int DebtID { get; set; }
         public string? Comments { get; set; }
     }
 }
